chore(docs): move mobile patterns into patterns/mobile folder and adjust import paths
